Finalize Task 1 of Assignment 3 - Mutation Testing

See merge request cse2115/2022-2023/SEM22a!39

+++ voting-microservice/src/test/java/voting/models/ProposalModelTest.java
@@ -16,7 +16,7 @@ class ProposalModelTest {
 
     @ParameterizedTest
     @MethodSource("argGen")
-    void isValidTest(int hoaId, String name, String desc, TimeModel scheduledFor, boolean expected, String testdesc) {
+    void isValidTest(long hoaId, String name, String desc, TimeModel scheduledFor, boolean expected, String testdesc) {
         ProposalModel sut = new ProposalModel(name, desc, hoaId, scheduledFor);
         assertEquals(expected, sut.isValid(), testdesc);
     }
